+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/ImpTest.hs
@@ -16,11 +16,13 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE PatternSynonyms #-}
 {-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TupleSections #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilyDependencies #-}
 {-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE UndecidableInstances #-}
 {-# LANGUAGE UndecidableSuperClasses #-}
+{-# LANGUAGE ViewPatterns #-}
 
 module Test.Cardano.Ledger.Shelley.ImpTest (
   ImpTestM,
@@ -55,7 +57,12 @@ module Test.Cardano.Ledger.Shelley.ImpTest (
   submitTxAnn_,
   submitFailingTx,
   submitFailingTxM,
+  submitBlock_,
+  submitBlock,
+  submitFailingBlock,
+  submitFailingBlockM,
   trySubmitTx,
+  trySubmitBlock,
   impShelleyExpectTxSuccess,
   modifyNES,
   getProtVer,
@@ -145,6 +152,12 @@ module Test.Cardano.Ledger.Shelley.ImpTest (
   withNoFixup,
   withPostFixup,
   withPreFixup,
+  withTxsInBlock_,
+  withTxsInBlock,
+  withTxsInFailingBlock,
+  withTxsInFailingBlockM,
+  impTxsFrom,
+  impSTSEventsFrom,
   impNESL,
   impGlobalsL,
   impLastTickG,
@@ -167,10 +180,10 @@ import Cardano.Ledger.Address (
   RewardAccount (..),
   bootstrapKeyHash,
  )
-import Cardano.Ledger.BHeaderView (BHeaderView)
+import Cardano.Ledger.BHeaderView (BHeaderView (..))
 import Cardano.Ledger.BaseTypes
 import Cardano.Ledger.Binary (DecCBOR, EncCBOR)
-import Cardano.Ledger.Block (Block)
+import Cardano.Ledger.Block (Block (..))
 import Cardano.Ledger.Coin
 import Cardano.Ledger.Compactible (fromCompact)
 import Cardano.Ledger.Credential (Credential (..), Ptr, StakeReference (..), credToText)
@@ -209,7 +222,7 @@ import Cardano.Ledger.Shelley.LedgerState (
 import Cardano.Ledger.Shelley.Rules (
   BbodyEnv (..),
   LedgerEnv (..),
-  ShelleyBbodyState,
+  ShelleyBbodyState (..),
   ShelleyDelegPredFailure,
   ShelleyPoolPredFailure,
   ShelleyUtxoPredFailure,
@@ -234,7 +247,7 @@ import Cardano.Ledger.TxIn (TxId (..), TxIn (..))
 import Cardano.Ledger.Val (Val (..))
 import Cardano.Slotting.EpochInfo (fixedEpochInfo)
 import Cardano.Slotting.Time (mkSlotLength)
-import Control.Monad (forM)
+import Control.Monad (forM, (<=<))
 import Control.Monad.IO.Class
 import Control.Monad.Reader (MonadReader (..), asks)
 import Control.Monad.State.Strict (MonadState (..), evalStateT, get, gets, modify, put)
@@ -249,6 +262,7 @@ import Control.State.Transition.Extended (
   ValidationPolicy (..),
  )
 import Data.Bifunctor (first)
+import Data.Bool (bool)
 import Data.Coerce (coerce)
 import Data.Data (Proxy (..), type (:~:) (..))
 import Data.Default (Default (..))
@@ -260,6 +274,7 @@ import Data.Map.Strict (Map)
 import qualified Data.Map.Strict as Map
 import Data.Maybe (catMaybes, isNothing, mapMaybe)
 import Data.Ratio ((%))
+import qualified Data.Sequence as Seq
 import Data.Sequence.Strict (StrictSeq (..))
 import qualified Data.Sequence.Strict as SSeq
 import qualified Data.Set as Set
@@ -344,7 +359,8 @@ data ImpTestState era = ImpTestState
   , impNativeScripts :: !(Map ScriptHash (NativeScript era))
   , impLastTick :: !SlotNo
   , impGlobals :: !Globals
-  , impEvents :: [SomeSTSEvent era]
+  , impRecordTxs :: Bool
+  , impEvents :: (Seq.Seq (Tx TopTx era), Seq.Seq (SomeSTSEvent era))
   }
 
 -- | This is a preliminary state that is used to prepare the actual `ImpTestState`
@@ -407,18 +423,33 @@ impNativeScriptsG ::
   SimpleGetter (ImpTestState era) (Map ScriptHash (NativeScript era))
 impNativeScriptsG = impNativeScriptsL
 
-impEventsL :: Lens' (ImpTestState era) [SomeSTSEvent era]
+impRecordTxsL :: Lens' (ImpTestState era) Bool
+impRecordTxsL = lens impRecordTxs (\x y -> x {impRecordTxs = y})
+
+impEventsL :: Lens' (ImpTestState era) (Seq.Seq (Tx TopTx era), Seq.Seq (SomeSTSEvent era))
 impEventsL = lens impEvents (\x y -> x {impEvents = y})
 
 class
   ( ShelleyEraTest era
-  , -- For BBODY rule
+  , -- For the BBODY rule
     STS (EraRule "BBODY" era)
   , BaseM (EraRule "BBODY" era) ~ ShelleyBase
   , Environment (EraRule "BBODY" era) ~ BbodyEnv era
   , State (EraRule "BBODY" era) ~ ShelleyBbodyState era
   , Signal (EraRule "BBODY" era) ~ Block BHeaderView era
+  , Eq (Event (EraRule "BBODY" era))
   , ToExpr (Event (EraRule "BBODY" era))
+  , Typeable (Event (EraRule "BBODY" era))
+  , NFData (BlockBody era)
+  , ToExpr (BlockBody era)
+  , NFData (PredicateFailure (EraRule "BBODY" era))
+  , ToExpr (PredicateFailure (EraRule "BBODY" era))
+  , EncCBOR (PredicateFailure (EraRule "BBODY" era))
+  , DecCBOR (PredicateFailure (EraRule "BBODY" era))
+  , -- For the LEDGERS rule
+    Eq (Event (EraRule "LEDGERS" era))
+  , ToExpr (Event (EraRule "LEDGERS" era))
+  , Typeable (Event (EraRule "LEDGERS" era))
   , State (EraRule "LEDGERS" era) ~ LedgerState era
   , -- For the LEDGER rule
     STS (EraRule "LEDGER" era)
@@ -641,6 +672,7 @@ defaultInitImpTestState nes = do
       , impNativeScripts = mempty
       , impLastTick = slotNo
       , impGlobals = globals
+      , impRecordTxs = False
       , impEvents = mempty
       }
 
@@ -652,7 +684,7 @@ withEachEraVersion ::
 withEachEraVersion specWith =
   withImpInit @(LedgerSpec era) $ do
     forM_ (eraProtVersions @era) $ \protVer ->
-      describe (show protVer) $
+      describe ("Protocol " <> show protVer) $
         modifyImpInitProtVer protVer specWith
 
 shelleyModifyImpInitProtVer ::
@@ -895,7 +927,7 @@ itePostEpochBoundaryHookL ::
     )
 itePostEpochBoundaryHookL = lens itePostEpochBoundaryHook (\x y -> x {itePostEpochBoundaryHook = y})
 
-instance MonadWriter [SomeSTSEvent era] (ImpTestM era) where
+instance MonadWriter (Seq.Seq (Tx TopTx era), Seq.Seq (SomeSTSEvent era)) (ImpTestM era) where
   writer (x, evs) = (impEventsL %= (<> evs)) $> x
   listen act = do
     oldEvs <- use impEventsL
@@ -907,6 +939,22 @@ instance MonadWriter [SomeSTSEvent era] (ImpTestM era) where
   pass act = do
     ((a, f), evs) <- listen act
     writer (a, f evs)
+
+impEventsFrom ::
+  ImpTestM era () ->
+  ImpTestM era (Seq.Seq (Tx TopTx era), Seq.Seq (SomeSTSEvent era))
+impEventsFrom = fmap snd . listen
+
+impTxsFrom :: ImpTestM era () -> ImpTestM era [Tx TopTx era]
+impTxsFrom act = do
+  prev <- use impRecordTxsL
+  impRecordTxsL .= True
+  txs <- toList . fst <$> impEventsFrom act
+  impRecordTxsL .= prev
+  pure txs
+
+impSTSEventsFrom :: ImpTestM era () -> ImpTestM era [SomeSTSEvent era]
+impSTSEventsFrom = fmap (toList . snd) . impEventsFrom
 
 runShelleyBase :: ShelleyBase a -> ImpTestM era a
 runShelleyBase act = do
@@ -1187,12 +1235,11 @@ trySubmitTx tx = do
   logToExpr txFixed
   st <- gets impNES
   lEnv <- impLedgerEnv st
-  ImpTestState {impRootTxIn} <- get
   res <- tryRunImpRule @"LEDGER" lEnv (st ^. nesEsL . esLStateL) txFixed
-  globals <- use impGlobalsL
-  let trc = TRC (lEnv, st ^. nesEsL . esLStateL, txFixed)
 
   -- Check for conformance
+  globals <- use impGlobalsL
+  let trc = TRC (lEnv, st ^. nesEsL . esLStateL, txFixed)
   asks itePostSubmitTxHook >>= (\f -> f globals trc res)
 
   case res of
@@ -1200,25 +1247,32 @@ trySubmitTx tx = do
       -- Verify that produced predicate failures are ready for the node-to-client protocol
       liftIO $ forM_ predFailures $ roundTripEraExpectation @era
       pure $ Left (predFailures, txFixed)
-    Right (st', events) -> do
-      let txId = TxId . hashAnnotated $ txFixed ^. bodyTxL
-          outsSize = SSeq.length $ txFixed ^. bodyTxL . outputsTxBodyL
-          rootIndex
-            | outsSize > 0 = outsSize - 1
-            | otherwise = error ("Expected at least 1 output after submitting tx: " <> show txId)
-      tell $ fmap (SomeSTSEvent @era @"LEDGER") events
-      modify $ impNESL . nesEsL . esLStateL .~ st'
+    Right (newState, events) -> do
+      impNESL . nesEsL . esLStateL .= newState
+
+      txsToRecord <- bool mempty (Seq.singleton txFixed) <$> use impRecordTxsL
+      tell (txsToRecord, Seq.fromList $ SomeSTSEvent @era @"LEDGER" <$> events)
+
+      ImpTestState {impRootTxIn} <- get
       UTxO utxo <- getUTxO
-      -- This TxIn is in the utxo, and thus can be the new root, only if the transaction
-      -- was phase2-valid.  Otherwise, no utxo with this id would have been created, and
-      -- so we need to set the new root to what it was before the submission.
-      let assumedNewRoot = TxIn txId (mkTxIxPartial (fromIntegral rootIndex))
-      let newRoot
-            | Map.member assumedNewRoot utxo = assumedNewRoot
-            | Map.member impRootTxIn utxo = impRootTxIn
-            | otherwise = error "Root not found in UTxO"
+      let
+        txId = TxId . hashAnnotated $ txFixed ^. bodyTxL
+        outsSize = SSeq.length $ txFixed ^. bodyTxL . outputsTxBodyL
+        rootIndex
+          | outsSize > 0 = outsSize - 1
+          | otherwise = error ("Expected at least 1 output after submitting tx: " <> show txId)
+        -- This TxIn is in the utxo, and thus can be the new root, only if the transaction
+        -- was phase2-valid.  Otherwise, no utxo with this id would have been created, and
+        -- so we need to set the new root to what it was before the submission.
+        assumedNewRoot = TxIn txId (mkTxIxPartial (fromIntegral rootIndex))
+        newRoot
+          | Map.member assumedNewRoot utxo = assumedNewRoot
+          | Map.member impRootTxIn utxo = impRootTxIn
+          | otherwise = error "Root not found in UTxO"
       impRootTxInL .= newRoot
+
       expectTxSuccess txFixed
+
       pure $ Right txFixed
 
 -- | Submit a transaction that is expected to be rejected with the given predicate failures.
@@ -1246,6 +1300,168 @@ submitFailingTxM tx mkExpectedFailures = do
   (predFailures, fixedUpTx) <- expectLeftDeepExpr =<< trySubmitTx tx
   expectedFailures <- mkExpectedFailures fixedUpTx
   predFailures `shouldBeExpr` expectedFailures
+
+-- | Gather all the txs submitted by @act@ and resubmit them as a block that's expected to succeed.
+withTxsInBlock_ ::
+  ( HasCallStack
+  , ShelleyEraImp era
+  ) =>
+  ImpTestM era a ->
+  ImpTestM era ()
+withTxsInBlock_ = void . withTxsInBlock . void
+
+-- | Gather all the txs submitted by @act@ and resubmit them as a block that's expected to succeed.
+withTxsInBlock ::
+  ( HasCallStack
+  , ShelleyEraImp era
+  ) =>
+  ImpTestM era () ->
+  ImpTestM era (Block BHeaderView era)
+withTxsInBlock act = do
+  txs <- simulateThenRestore $ impTxsFrom act
+  submitBlock txs
+
+-- | Gather all the txs submitted by @act@ and resubmit them as a block
+-- that's expected to fail with the given predicate failures.
+withTxsInFailingBlock ::
+  ( HasCallStack
+  , ShelleyEraImp era
+  ) =>
+  ImpTestM era () ->
+  NonEmpty (PredicateFailure (EraRule "BBODY" era)) ->
+  ImpTestM era ()
+withTxsInFailingBlock act = withTxsInFailingBlockM act . const . pure
+
+-- | Gather all the txs submitted by @act@ and resubmit them as a block that's expected to fail
+-- and compute the expected predicate failures from the created block using the supplied action.
+withTxsInFailingBlockM ::
+  ( HasCallStack
+  , ShelleyEraImp era
+  ) =>
+  ImpTestM era () ->
+  (Block BHeaderView era -> ImpTestM era (NonEmpty (PredicateFailure (EraRule "BBODY" era)))) ->
+  ImpTestM era ()
+withTxsInFailingBlockM act mkExpectedFailures = do
+  txs <- simulateThenRestore $ impTxsFrom act
+  submitFailingBlockM txs mkExpectedFailures
+
+-- | Submit a list of transactions as a block that's expected to succeed.
+-- The transactions are not modified to make them balance.
+submitBlock_ ::
+  ( HasCallStack
+  , ShelleyEraImp era
+  ) =>
+  [Tx TopTx era] -> ImpTestM era ()
+submitBlock_ = void . submitBlock
+
+-- | Submit a list of transactions as a block that's expected to succeed.
+-- The transactions are not modified to make them balance.
+submitBlock ::
+  ( HasCallStack
+  , ShelleyEraImp era
+  ) =>
+  [Tx TopTx era] -> ImpTestM era (Block BHeaderView era)
+submitBlock = expectRightDeepExpr . first fst <=< trySubmitBlock
+
+-- | Submit a list of transactions as a block that's expected to fail
+-- with the given predicate failures.
+-- The transactions are not modified to make them balance.
+submitFailingBlock ::
+  ( HasCallStack
+  , ShelleyEraImp era
+  ) =>
+  [Tx TopTx era] ->
+  NonEmpty (PredicateFailure (EraRule "BBODY" era)) ->
+  ImpTestM era ()
+submitFailingBlock txs = submitFailingBlockM txs . const . pure
+
+-- | Submit a list of transactions as a block that's expected to be rejected,
+-- and compute the expected predicate failures from the created block using the supplied action.
+-- The transactions are not modified to make them balance.
+submitFailingBlockM ::
+  ( HasCallStack
+  , ShelleyEraImp era
+  ) =>
+  [Tx TopTx era] ->
+  (Block BHeaderView era -> ImpTestM era (NonEmpty (PredicateFailure (EraRule "BBODY" era)))) ->
+  ImpTestM era ()
+submitFailingBlockM txs mkExpectedFailures = do
+  (predFailures, block) <- expectLeftDeepExpr =<< trySubmitBlock txs
+  expectedFailures <- mkExpectedFailures block
+  predFailures `shouldBeExpr` expectedFailures
+
+-- | Try to submit a list of transactions as a block,
+-- Return the block that was created using the transactions and any predicate
+-- failures that are produced.
+-- The transactions are not modified to make them balance.
+trySubmitBlock ::
+  forall era.
+  ShelleyEraImp era =>
+  [Tx TopTx era] ->
+  ImpTestM
+    era
+    ( Either
+        (NonEmpty (PredicateFailure (EraRule "BBODY" era)), Block BHeaderView era)
+        (Block BHeaderView era)
+    )
+trySubmitBlock (SSeq.fromList -> txs) = do
+  kh <- freshKeyHash
+  slotNo <- use impLastTickG
+  let blockHeader =
+        BHeaderView
+          { bhviewID = kh
+          , bhviewBSize = fromIntegral $ bBodySize (ProtVer (eraProtVerLow @era) 0) blockBody
+          , bhviewHSize = 0
+          , bhviewBHash = hashBlockBody blockBody
+          , bhviewSlot = slotNo
+          }
+      blockBody = mkBasicBlockBody @era & txSeqBlockBodyL .~ txs
+      block = Block {blockHeader, blockBody}
+
+  nes <- use impNESL
+  let ls = nes ^. nesEsL . esLStateL
+      pp = nes ^. nesEsL . curPParamsEpochStateL @era
+  res <-
+    tryRunImpRule @"BBODY"
+      (BbodyEnv pp (nes ^. chainAccountStateL))
+      (BbodyState ls (BlocksMade Map.empty))
+      block
+
+  case res of
+    Left predFailures -> do
+      -- Verify that produced predicate failures are ready for the node-to-client protocol
+      liftIO $ forM_ predFailures $ roundTripEraExpectation @era
+      pure $ Left (predFailures, block)
+    Right (BbodyState newState _, events) -> do
+      impNESL . nesEsL . esLStateL .= newState
+
+      txsToRecord <- bool mempty (SSeq.fromStrict txs) <$> use impRecordTxsL
+      tell (txsToRecord, Seq.fromList $ SomeSTSEvent @era @"BBODY" <$> events)
+
+      ImpTestState {impRootTxIn} <- get
+      UTxO utxo <- getUTxO
+
+      case txs of
+        _ SSeq.:|> txFinal -> do
+          let
+            txId = TxId . hashAnnotated $ txFinal ^. bodyTxL
+            outsSize = SSeq.length $ txFinal ^. bodyTxL . outputsTxBodyL
+            rootIndex
+              | outsSize > 0 = outsSize - 1
+              | otherwise = error ("Expected at least 1 output after submitting tx: " <> show txId)
+            -- This TxIn is in the utxo, and thus can be the new root, only if the transaction
+            -- was phase2-valid.  Otherwise, no utxo with this id would have been created, and
+            -- so we need to set the new root to what it was before the submission.
+            assumedNewRoot = TxIn txId (mkTxIxPartial (fromIntegral rootIndex))
+            newRoot
+              | Map.member assumedNewRoot utxo = assumedNewRoot
+              | Map.member impRootTxIn utxo = impRootTxIn
+              | otherwise = error "Root not found in UTxO"
+          impRootTxInL .= newRoot
+        SSeq.Empty ->
+          pure ()
+
+      pure $ Right block
 
 tryRunImpRule ::
   forall rule era.
@@ -1326,7 +1542,7 @@ runImpRule env st sig = do
           unlines $
             ("Failed to run " <> ruleName <> ":") : map show (toList fs)
       Right res -> evaluateDeep res
-  tell $ fmap (SomeSTSEvent @era @rule) ev
+  tell (mempty, Seq.fromList $ SomeSTSEvent @era @rule <$> ev)
   pure res
 
 -- | Runs the TICK rule once