+++ eras/shelley/impl/CHANGELOG.md
@@ -33,6 +33,20 @@
 
 ### `testlib`
 
+* Added:
+  - `submitBlock_`
+  - `submitBlock`
+  - `submitFailingBlock`
+  - `submitFailingBlockM`
+  - `trySubmitBlock`
+  - `withTxsInBlock_`
+  - `withTxsInBlock`
+  - `withTxsInFailingBlock`
+  - `withTxsInFailingBlockM`
+  - `impTxsFrom`
+  - `impSTSEventsFrom`
+* Changed type of `ImpTestState.impEvents` field
+* Added `ImpTestState.impRecordTxs` field
 * Rename `multisig_script` -> `shelley_native_script` in CDDL
 * Move `script_hash` to core CDDL
 * Move `transaction_metadatum_label`, `transaction_metadatum`, `transaction_index` and `metadata_hash` to core CDDL